chore(suite): make seach action component reusable
